Small refactoring + added settings for Calenday
